PySparkStreams/requirements.txt
# requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
pandas==2.1.0
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3